Fix local theme building during development
- Themes need to be in the shared volume for the live reload plugin to work
- We don't want to always overwrite themes if we're doing local development

+++ entrypoint.sh
@@ -5,13 +5,16 @@ set -e
 # The base Ghost image will copy themes from the
 # image /content/themes folder on startup, but only
 # if they don't already exist.
-baseDir="$GHOST_SOURCE/content"
-for dir in "$baseDir"/themes/*/; do
-  rm -rf "$GHOST_CONTENT/${dir#$baseDir/}"
-done
+if [ "$GHOST_REDEPLOY_THEMES" = "true" ]; then
+  baseDir="$GHOST_SOURCE/content"
+  for dir in "$baseDir"/themes/*/; do
+    rm -rf "$GHOST_CONTENT/${dir#$baseDir/}"
+  done
+fi
 
 # We also want our configuration to be overwritten with
-# any new values baked into the image.
+# any new values baked into the image.  Removing src/config.js
+# forces any
 rm -f "$GHOST_SOURCE"/config.js
 cp -f "$GHOST_SOURCE"/config.override.js "$GHOST_CONTENT"/config.js
